moved files from the UI folder to User  Post folder
