+++ packages/nextjs/src/components/quickPost/UserProfilePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import UserProfilePostsGrid from './post-grid/userprofilepost';
 
-// Placeholder for the user data component that would be above the posts grid
+
 const UserProfileHeader = () => (
   <div className="mb-8 p-6 bg-white dark:bg-gray-800 rounded-lg shadow">
     <div className="flex flex-col md:flex-row items-start md:items-center gap-6">
